refactor [#86] DeleteAlert, InputAlert을 barrel index file에서 export
