+++ src/components/modal/index.ts
@@ -1,3 +1,5 @@
 export { default as Alert } from "./Alert";
+export { default as DeleteAlert } from "./DeleteAlert";
+export { default as InputAlert } from "./InputAlert";
 export { default as Modal } from "./Modal";
 export { default as Sheet } from "./Sheet";
